Rebase October stake changes to 2.5MM BSQ
This change reflects the new genesis distribution at [1]. Numbers were
generated with the following commands:

    cut -d, -f4 stake.csv \
        | awk '{n=($1*(2500000/2548500)); printf("%0.2f\n", n)}' \
        | awk '{n+=$1}END{printf("%0.2f\n", n)}' \
        > rebased-stake.csv;
        mv rebased-stake.csv stake.csv;

where the number 2548500 above was the amount of BSQ *after* October
voting was complete, i.e. the sum of the 4th column of the stake.csv
file prior to running the command above.

Note that 2-decimal precision was used here in order to have the rebased
value sum to exactly 2,500,000.00 BSQ. The spreadsheet calculation that
led up to the redistribution transaction linked below did not have this
precision, and due to rounding caused the resulting amount of BSQ to
equal 2,499,986 BSQ, which is why the last 14 addresses had 1 BSQ added
to them.

The 2-decimal precision numbers in this change are what have been fed
into the BSQ stake and voting spreadsheet at [2] (see the 2017.11
worksheet). Even though these numbers are slightly different than what's
represented on the testnet blockchain, we'll use them going forward into
next month's redistribution transaction, such that we get accurate
numbers going forward. We'll probably have to work with 2-decimal
precision in everything we do from this point on.

See bisq-network/dao#9

[1]: https://testnet.smartbit.com.au/tx/f8b65c65624bd822f92480c39959f8ae4a6f94a9841c1625464ec6353cfba1d9
[2]: https://docs.google.com/spreadsheets/d/1xlXDswj3251BPCOcII-UyWlX7o7jMkfYBE-IZ5te5Ck/edit#gid=1776326737